+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 apidoc extension Change Log
 3.0.7 under development
 -----------------------
 
-- no changes in this release.
+- Bug #293: Logged errors and warnings were duplicated in some cases (rhertogh)
 
 
 3.0.6 November 18, 2022